refactor(events): split EssentialsStep under the line cap (#544)
EssentialsStep.svelte was 802 lines. Extract the self-contained slug editor —
its own edit-mode state machine, slug-update mutation (own accessToken derive),
format validation and URL preview — into EssentialsStepSlugEditor.svelte (flat
prefixed sibling). Props: eventId, eventSlug, organizationSlug, onSlugUpdated.
The `{#if isEditMode && eventId && organizationSlug}` gate stays in the parent
so the child only mounts in edit mode, exactly as before.

The is_open_ended open-ended toggle + end-date hide logic (PR #495) stays in
EssentialsStep, untouched. EssentialsStep 802 -> 611. Behavior unchanged.
